ModelParameters

Types

ModelParameters.AbstractParameter — Type

Abstract supertype for parameters. Currently, the subtypes are Parameters and BoundedParameter.

ModelParameters.AbstractParameterSet — Type

Abstract supertype for different definitions of parameter sets. Currently the subtypes are ParameterGroup and ParameterSpace.

ModelParameters.BoundedParameter — Type

A simple wrapper that contains the value of the parameter, the lower and upper bounds, and a description.

ModelParameters.Parameter — Type

A simple wrapper that contains the value of the parameter and a description.

ModelParameters.ParameterGroup — Type

Abstract supertype for structs that contain multiple Parameters and/or BoundedParameters. Useful for printing a single struct that holds parameters.

ModelParameters.ParameterSpace — Type

Abstract supertype that contains all the parameters in the model. Useful for printing the parameter space.

Functions

Base.getproperty — Method

Base.getproperty(name::ParameterGroup, val::Symbol)

Extension of getproperty function in order to call the value of an AbstractParameter within a ParameterGroup as T.x instead of T.x.value

Base.show — Method

Extension of the Base.show function in order to print all the ParameterGroups, including the description in a readable format.

ModelParameters.:← — Method

(←)(str, tpl::Tuple)

Substitutes the value of a single or multiple fields of struct str. The tuple has the form (field::Symbol, new_value). Returns the modified str.

ModelParameters.description — Method

A function that returns the description of an AbstractParameter within a ParameterGroup. This is necessary because of the definition of getproperty for ParameterGroup that returns the value of the parameter instead of AbstractParameter. For example, T.x returns x.value and not x, which contains the description in the wrapper.

ModelParameters.get_bounds — Method

get_bounds(name::ParameterGroup, val::Symbol)

Returns the lower and upper bound of a single parameter withing a ParameterGroup. Useful for constrained optimization.

ModelParameters.get_initial_guess — Method

get_initial_guess(parameters)

Returns the values of parameter vector. Useful for initial guess for optimization.

ModelParameters.get_lower_bounds — Method

get_lower_bounds(parameters::Array{Float64)

Returns the lower bounds of parameter vector. Useful for constrained optimization.

ModelParameters.get_upper_bounds — Method

get_upper_bounds(parameters::Array{Float64)

Returns the upper bounds of parameter vector. Useful for constrained optimization.

ModelParameters.set — Method

set(str,var::Symbol,val)

Substitutes the value a single field var with val. Returns the modified str.